itemservice2/src/main/java/hello/itemservice2/domain/member/MemberService.java
package hello.itemservice2.domain.member;

import java.util.List;
import java.util.Optional;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class MemberService {

  private final MemberRepository memberRepository;

  @Autowired
  public MemberService(MemberRepository memberRepository) {
    this.memberRepository = memberRepository;
  }

  /**
   *  회원 가입
   */
  public Long join(Member member) {
    validateDuplicateMember(member);
    memberRepository.save(member);
    return member.getId();
  }

  private void validateDuplicateMember(Member member) {
    memberRepository.findByLoginId(member.getLoginId())
                    .ifPresent( m -> {
                      throw new IllegalStateException("이미 존재하는 회원입니다.");
                    }); //TODO 500 에러 말고 다른 방법으로 에러 처리 하기
  }

  /**
   * 전체 회원 조회
   */
  public List<Member> findMembers() {
    return memberRepository.findAll();
  }

  public Optional<Member> findByLoginId(String loginId) {
    return memberRepository.findByLoginId(loginId);
  }

}
